[Phyo_branch] Continuing with centralizing code (#57)
* Centralized Navigation bar

No more duplicated code for <head> and navigation bar. Used Jinja2 which we can find the common code under base.html.

* Update main.db

Unable to switch to main branch till this file is committed.

* Forgetpassword accessible  while logged in

the forget password page is no longer accessible while the user is logged in

* PIN error under changing password and forget password accessible issue

Now can change the password only when correct PIN is provided. (Tested and work)

forget password password is not longer accessible while logged in. (Tested and worked)

* Continuing with centralizing code

* now both /profile and /profile/userid is working for navi
* index, registerm register_complete and signs 's navi are updated and centralized.
* Have more pages to do

* only digital PIN is allowed