[Phyo branch] Centralized Help navigation (#69)
* Centralized Navigation bar

No more duplicated code for <head> and navigation bar. Used Jinja2 which we can find the common code under base.html.

* Update main.db

Unable to switch to main branch till this file is committed.

* Forgetpassword accessible  while logged in

the forget password page is no longer accessible while the user is logged in

* PIN error under changing password and forget password accessible issue

Now can change the password only when correct PIN is provided. (Tested and work)

forget password password is not longer accessible while logged in. (Tested and worked)

* Continuing with centralizing code

* now both /profile and /profile/userid is working for navi
* index, registerm register_complete and signs 's navi are updated and centralized.
* Have more pages to do

* only digital PIN is allowed

* Update main.db

* Validation of Login, Register and Forgot password

Changed the repeated validation to only one. Mainly using html built-in validation.

Improved the error msg sentences.

Changed the unmeaningful id's names accordingly.

* Registration error style, Check email before changing email, definition of password validation

Registration page error style is changed to highlight the text with bold and red style. (instead of id, used class)

While changing email, the email can only be updated, if the new email is not existed in the system.

Validation of modifycenter are changed
* JS for PIN is no longer checked password instead its PIN now
* definition of password is changed.
* modifycenter's nav is now centeralized.

* Centralize code for common code - Part 2

Below templates are completed for this commit and update base.html for Help tab.
* accept_request.html
* answerrequest.html
* chat_details.html
* confirm_buy.html (/confirmpayment)
* helpcenter.html (Error)
* myrequest.html
* newchat.html
* newrequest.html
* newthread.html
* requestDetail.html
* search_result.html
* thread_details.html (/thread)

* Chat Convo alert is removed

Chat Convo alert is removed meaning no more alert when you reply in the chatroom

* Delete main.db as per Phyo's request

* Delete main.db & helpcenter.html as per Phyo's request

* Add old helpcenter.html

* Centralized help nav

---------

Co-authored-by: xosadmin <9164356+xosadmin@users.noreply.github.com>